refactor(web): extract use-node-boxes.ts from SpatialEditor
The pre-scene node-geometry memos (boxes, selectedBox, selectedNode)
move verbatim into their own hook, called exactly where they stood —
before useViewportControls consumes `boxes`, and before the scene the
worker lays out later depends on nothing here. First of six clusters
splitting SpatialEditor.tsx's derived-scene pipeline.
